Implement Some Very Nice Waits
Additionally, everything works again. Turns out I was logged into the AUT in my main instance of Firefox which seems to share its LocalStorage with the WebDriver client, so Selenium couldn't find the login button because I was already logged in and there was no login button to find. So stupid!